+++ src/store/actions/profileActions.js
@@ -1,0 +1,3 @@
+import { PROFILE_HANDLECHANGE } from '../types'
+
+export const handleChange = (event) => ({type: PROFILE_HANDLECHANGE, payload: event.target})
