--- src/store/selectors/chats.js
@@ -1,4 +1,0 @@
-export const chatsKeysSelector = state => Object.keys(state.chats)
-export const chatsSelector = state => state.chats
-export const chatsListInit = state => chatsKeysSelector.map(key => ({id: key, name: state.chats[key].name}))
-// export const chatsListInit = state => Object.keys(state.chats).map(key => ({id: key, name: state.chats[key].name}))
