+++ src/store/selectors/chatsSelectors.js
@@ -1,0 +1,3 @@
+export const chatsKeysSelector = state => Object.keys(state.chats)
+export const chatsSelector = state => state.chats
+export const chatsListInit = state => Object.keys(state.chats).map(key => ({id: key, name: state.chats[key].name}))
